6-minute sessions, a 300-item decor tab, and a cleaner mobile top bar (#21)
- Quiz sessions now grant 6 minutes of play (was 4)
- New 'Objets deco' inventory tab: 300 generated decorative blocks
  (10 patterns - uni, briques, planches, damier, pois, rayures, lignes,
  zigzag, cadre, losange - x 30 named colors), paginated 60 per page
  with prev/next controls; picking one assigns it to the hotbar slot
  like any block, and the minimap colors them correctly
- Chunk storage widened to Uint16Array since block ids now exceed 255;
  texture atlas grown to 20x17 tiles
- Mobile top bar fix: the session timer moves to its own row under the
  buttons so it can never overlap the backpack/dex/pause buttons again
- Service worker cache bumped to v14

+++ src/world.js
@@ -227,7 +227,8 @@ export class World {
   }
 
   generateChunk(cx, cz) {
-    const data = new Uint8Array(CHUNK * CHUNK * HEIGHT);
+    // 16-bit: block ids go beyond 255 with the decorative set
+    const data = new Uint16Array(CHUNK * CHUNK * HEIGHT);
     const baseX = cx * CHUNK, baseZ = cz * CHUNK;
 
     for (let z = 0; z < CHUNK; z++) {
